+++ tests/testthat/helper-queue.R
@@ -7,3 +7,14 @@ test_redis_available <- function() {
 }
 
 Sys.unsetenv("HINTR_QUEUE_ID")
+
+MockQueue <- R6::R6Class(
+  "MockQueue",
+  inherit = Queue,
+  cloneable = FALSE,
+  public = list(
+    submit = function(data, options) {
+      self$queue$enqueue_(quote(stop("test error")))
+    }
+  )
+)