+++ tests/testthat/test-server.R
@@ -150,6 +150,7 @@ test_that("validate baseline", {
 })
 
 test_that("model interactions", {
+  test_mock_model_available()
   server <- hintr_server()
 
   ## Submit a model run
@@ -164,17 +165,19 @@ test_that("model interactions", {
   expect_equal(names(response$data), c("id"))
 
   ## Get the status
-  Sys.sleep(2)
-  r <- httr::GET(paste0(server$url, "/model/status/", response$data$id))
-  expect_equal(httr::status_code(r), 200)
-  response <- response_from_json(r)
-  expect_equal(response$status, "success")
-  expect_equal(response$errors, list())
-  expect_equal(response$data$done, TRUE)
-  expect_equal(response$data$status, "COMPLETE")
-  expect_equal(response$data$success, TRUE)
-  expect_equal(response$data$queue, 0)
-  expect_true("id" %in% names(response$data))
+  testthat::try_again(4, {
+    Sys.sleep(2)
+    r <- httr::GET(paste0(server$url, "/model/status/", response$data$id))
+    expect_equal(httr::status_code(r), 200)
+    response <- response_from_json(r)
+    expect_equal(response$status, "success")
+    expect_equal(response$errors, list())
+    expect_equal(response$data$done, TRUE)
+    expect_equal(response$data$status, "COMPLETE")
+    expect_equal(response$data$success, TRUE)
+    expect_equal(response$data$queue, 0)
+    expect_true("id" %in% names(response$data))
+  })
 
   ## Get the result
   r <- httr::GET(paste0(server$url, "/model/result/", response$data$id))
@@ -341,19 +344,56 @@ test_that("worker information is returned", {
 })
 
 test_that("spectrum file download streams bytes", {
+  test_mock_model_available()
   server <- hintr_server()
-  r <- httr::GET(paste0(server$url, "/download/spectrum/", "id123"))
+
+  ## Run a model
+  submit <- file.path("payload", "model_submit_payload.json")
+  r <- httr::POST(paste0(server$url, "/model/submit"),
+                  body = httr::upload_file(submit),
+                  encode = "json")
   expect_equal(httr::status_code(r), 200)
-  expect_equal(httr::headers(r)$`content-type`, "application/octet-stream")
-  expect_equal(httr::headers(r)$`content-length`,
-               as.character(file.size(system_file("output", "malawi.zip"))))
+  response <- response_from_json(r)
+  expect_equal(response$status, "success")
+  expect_equal(response$errors, list())
+  expect_equal(names(response$data), c("id"))
+
+  ## Get the download
+  testthat::try_again(4, {
+    Sys.sleep(2)
+    r <- httr::GET(paste0(server$url, "/download/spectrum/", response$data$id))
+    expect_equal(httr::status_code(r), 200)
+    expect_equal(httr::headers(r)$`content-type`, "application/octet-stream")
+    expect_equal(
+      httr::headers(r)$`content-length`,
+      as.character(file.size(system_file("output", "malawi_spectrum_download.zip"))))
+  })
 })
 
-test_that("indicators file download streams bytes", {
+test_that("summary file download streams bytes", {
+  test_mock_model_available()
   server <- hintr_server()
-  r <- httr::GET(paste0(server$url, "/download/indicators/", "id123"))
+
+  ## Run a model
+  submit <- file.path("payload", "model_submit_payload.json")
+  r <- httr::POST(paste0(server$url, "/model/submit"),
+                  body = httr::upload_file(submit),
+                  encode = "json")
   expect_equal(httr::status_code(r), 200)
-  expect_equal(httr::headers(r)$`content-type`, "application/octet-stream")
-  expect_equal(httr::headers(r)$`content-length`,
-               as.character(file.size(system_file("output", "malawi.zip"))))
+  response <- response_from_json(r)
+  expect_equal(response$status, "success")
+  expect_equal(response$errors, list())
+  expect_equal(names(response$data), c("id"))
+
+  ## Get the download
+  testthat::try_again(4, {
+    Sys.sleep(2)
+    r <- httr::GET(paste0(server$url, "/download/summary/", response$data$id))
+    expect_equal(httr::status_code(r), 200)
+    expect_equal(httr::headers(r)$`content-type`, "application/octet-stream")
+    expect_equal(
+      httr::headers(r)$`content-length`,
+      as.character(file.size(system_file("output",
+                                         "malawi_summary_download.zip"))))
+  })
 })